@@ -4,8 +4,8 @@ default-members = ["crates/typst-cli"]
 resolver = "2"
 
 [workspace.package]
-version = "0.11.0"
-rust-version = "1.77" # also change in ci.yml
+version = "0.12.0"
+rust-version = "1.80" # also change in ci.yml
 authors = ["The Typst Project Developers"]
 edition = "2021"
 homepage = "https://typst.app"
