@@ -1,11 +1,10 @@
-use typst::diag::StrResult;
 use typst::text::FontVariant;
 use typst_kit::fonts::Fonts;
 
 use crate::args::FontsCommand;
 
 /// Execute a font listing command.
-pub fn fonts(command: &FontsCommand) -> StrResult<()> {
+pub fn fonts(command: &FontsCommand) {
     let fonts = Fonts::searcher()
         .include_system_fonts(!command.font_args.ignore_system_fonts)
         .search_with(&command.font_args.font_paths);
@@ -19,6 +18,4 @@ pub fn fonts(command: &FontsCommand) -> StrResult<()> {
             }
         }
     }
-
-    Ok(())
 }
