@@ -191,7 +191,7 @@ mod tests {
         // Set page width to 120pt with 10pt margins, so that the inner page is
         // exactly 100pt wide. Page height is unbounded and font size is 10pt so
         // that it multiplies to nice round numbers.
-        let mut lib = Library::default();
+        let mut lib = typst::Library::default();
         lib.styles
             .set(PageElem::set_width(Smart::Custom(Abs::pt(120.0).into())));
         lib.styles.set(PageElem::set_height(Smart::Auto));
