@@ -5,26 +5,32 @@
 //! See also [`pdf_writer::Content`].
 
 use ecow::eco_format;
-use pdf_writer::{
-    types::{ColorSpaceOperand, LineCapStyle, LineJoinStyle, TextRenderingMode},
-    Content, Finish, Name, Rect, Str,
+use pdf_writer::types::{
+    ColorSpaceOperand, LineCapStyle, LineJoinStyle, TextRenderingMode,
 };
-use typst::layout::{
+use pdf_writer::writers::PositionedItems;
+use pdf_writer::{Content, Finish, Name, Rect, Str};
+use typst_library::diag::{bail, error, SourceDiagnostic, SourceResult};
+use typst_library::foundations::Repr;
+use typst_library::layout::{
     Abs, Em, Frame, FrameItem, GroupItem, Point, Ratio, Size, Transform,
 };
-use typst::model::Destination;
-use typst::text::{color::is_color_glyph, Font, TextItem, TextItemView};
-use typst::utils::{Deferred, Numeric, SliceExt};
-use typst::visualize::{
+use typst_library::model::Destination;
+use typst_library::text::color::should_outline;
+use typst_library::text::{Font, Glyph, TextItem, TextItemView};
+use typst_library::visualize::{
     FillRule, FixedStroke, Geometry, Image, LineCap, LineJoin, Paint, Path, PathItem,
     Shape,
 };
+use typst_syntax::Span;
+use typst_utils::{Deferred, Numeric, SliceExt};
 
+use crate::color::PaintEncode;
 use crate::color_font::ColorFontMap;
 use crate::extg::ExtGState;
 use crate::image::deferred_image;
-use crate::{color::PaintEncode, resources::Resources};
-use crate::{deflate_deferred, AbsExt, EmExt};
+use crate::resources::Resources;
+use crate::{deflate_deferred, AbsExt, ContentExt, EmExt, PdfOptions, StrExt};
 
 /// Encode a [`Frame`] into a content stream.
 ///
@@ -35,13 +41,14 @@ use crate::{deflate_deferred, AbsExt, EmExt};
 ///
 /// [color glyph]: `crate::color_font`
 pub fn build(
+    options: &PdfOptions,
     resources: &mut Resources<()>,
     frame: &Frame,
     fill: Option<Paint>,
     color_glyph_width: Option<f32>,
-) -> Encoded {
+) -> SourceResult<Encoded> {
     let size = frame.size();
-    let mut ctx = Builder::new(resources, size);
+    let mut ctx = Builder::new(options, resources, size);
 
     if let Some(width) = color_glyph_width {
         ctx.content.start_color_glyph(width);
@@ -57,18 +64,18 @@ pub fn build(
 
     if let Some(fill) = fill {
         let shape = Geometry::Rect(frame.size()).filled(fill);
-        write_shape(&mut ctx, Point::zero(), &shape);
+        write_shape(&mut ctx, Point::zero(), &shape)?;
     }
 
     // Encode the frame into the content stream.
-    write_frame(&mut ctx, frame);
+    write_frame(&mut ctx, frame)?;
 
-    Encoded {
+    Ok(Encoded {
         size,
         content: deflate_deferred(ctx.content.finish()),
         uses_opacities: ctx.uses_opacities,
         links: ctx.links,
-    }
+    })
 }
 
 /// An encoded content stream.
@@ -91,6 +98,8 @@ pub struct Encoded {
 /// Content streams can be used for page contents, but also to describe color
 /// glyphs and patterns.
 pub struct Builder<'a, R = ()> {
+    /// Settings for PDF export.
+    pub(crate) options: &'a PdfOptions<'a>,
     /// A list of all resources that are used in the content stream.
     pub(crate) resources: &'a mut Resources<R>,
     /// The PDF content stream that is being built.
@@ -107,8 +116,13 @@ pub struct Builder<'a, R = ()> {
 
 impl<'a, R> Builder<'a, R> {
     /// Create a new content builder.
-    pub fn new(resources: &'a mut Resources<R>, size: Size) -> Self {
+    pub fn new(
+        options: &'a PdfOptions<'a>,
+        resources: &'a mut Resources<R>,
+        size: Size,
+    ) -> Self {
         Builder {
+            options,
             resources,
             uses_opacities: false,
             content: Content::new(),
@@ -187,9 +201,9 @@ pub(super) struct Transforms {
 }
 
 impl Builder<'_, ()> {
-    fn save_state(&mut self) {
+    fn save_state(&mut self) -> SourceResult<()> {
         self.saves.push(self.state.clone());
-        self.content.save_state();
+        self.content.save_state_checked()
     }
 
     fn restore_state(&mut self) {
@@ -267,13 +281,19 @@ impl Builder<'_, ()> {
         self.state.size = size;
     }
 
-    fn set_fill(&mut self, fill: &Paint, on_text: bool, transforms: Transforms) {
+    fn set_fill(
+        &mut self,
+        fill: &Paint,
+        on_text: bool,
+        transforms: Transforms,
+    ) -> SourceResult<()> {
         if self.state.fill.as_ref() != Some(fill)
             || matches!(self.state.fill, Some(Paint::Gradient(_)))
         {
-            fill.set_as_fill(self, on_text, transforms);
+            fill.set_as_fill(self, on_text, transforms)?;
             self.state.fill = Some(fill.clone());
         }
+        Ok(())
     }
 
     pub fn set_fill_color_space(&mut self, space: Name<'static>) {
@@ -292,7 +312,7 @@ impl Builder<'_, ()> {
         stroke: &FixedStroke,
         on_text: bool,
         transforms: Transforms,
-    ) {
+    ) -> SourceResult<()> {
         if self.state.stroke.as_ref() != Some(stroke)
             || matches!(
                 self.state.stroke.as_ref().map(|s| &s.paint),
@@ -300,7 +320,7 @@ impl Builder<'_, ()> {
             )
         {
             let FixedStroke { paint, thickness, cap, join, dash, miter_limit } = stroke;
-            paint.set_as_stroke(self, on_text, transforms);
+            paint.set_as_stroke(self, on_text, transforms)?;
 
             self.content.set_line_width(thickness.to_f32());
             if self.state.stroke.as_ref().map(|s| &s.cap) != Some(cap) {
@@ -324,6 +344,8 @@ impl Builder<'_, ()> {
             }
             self.state.stroke = Some(stroke.clone());
         }
+
+        Ok(())
     }
 
     pub fn set_stroke_color_space(&mut self, space: Name<'static>) {
@@ -346,26 +368,29 @@ impl Builder<'_, ()> {
 }
 
 /// Encode a frame into the content stream.
-pub(crate) fn write_frame(ctx: &mut Builder, frame: &Frame) {
+pub(crate) fn write_frame(ctx: &mut Builder, frame: &Frame) -> SourceResult<()> {
     for &(pos, ref item) in frame.items() {
         let x = pos.x.to_f32();
         let y = pos.y.to_f32();
         match item {
-            FrameItem::Group(group) => write_group(ctx, pos, group),
-            FrameItem::Text(text) => write_text(ctx, pos, text),
-            FrameItem::Shape(shape, _) => write_shape(ctx, pos, shape),
-            FrameItem::Image(image, size, _) => write_image(ctx, x, y, image, *size),
+            FrameItem::Group(group) => write_group(ctx, pos, group)?,
+            FrameItem::Text(text) => write_text(ctx, pos, text)?,
+            FrameItem::Shape(shape, _) => write_shape(ctx, pos, shape)?,
+            FrameItem::Image(image, size, span) => {
+                write_image(ctx, x, y, image, *size, *span)?
+            }
             FrameItem::Link(dest, size) => write_link(ctx, pos, dest, *size),
             FrameItem::Tag(_) => {}
         }
     }
+    Ok(())
 }
 
 /// Encode a group into the content stream.
-fn write_group(ctx: &mut Builder, pos: Point, group: &GroupItem) {
+fn write_group(ctx: &mut Builder, pos: Point, group: &GroupItem) -> SourceResult<()> {
     let translation = Transform::translate(pos.x, pos.y);
 
-    ctx.save_state();
+    ctx.save_state()?;
 
     if group.frame.kind().is_hard() {
         ctx.group_transform(
@@ -385,39 +410,35 @@ fn write_group(ctx: &mut Builder, pos: Point, group: &GroupItem) {
         ctx.content.end_path();
     }
 
-    write_frame(ctx, &group.frame);
+    write_frame(ctx, &group.frame)?;
     ctx.restore_state();
+
+    Ok(())
 }
 
 /// Encode a text run into the content stream.
-fn write_text(ctx: &mut Builder, pos: Point, text: &TextItem) {
-    let ttf = text.font.ttf();
-    let tables = ttf.tables();
-
-    // If the text run contains either only color glyphs (used for emojis for
-    // example) or normal text we can render it directly
-    let has_color_glyphs = tables.sbix.is_some()
-        || tables.cbdt.is_some()
-        || tables.svg.is_some()
-        || tables.colr.is_some();
-    if !has_color_glyphs {
-        write_normal_text(ctx, pos, TextItemView::all_of(text));
-        return;
+fn write_text(ctx: &mut Builder, pos: Point, text: &TextItem) -> SourceResult<()> {
+    if ctx.options.standards.pdfa && text.font.info().is_last_resort() {
+        bail!(
+            Span::find(text.glyphs.iter().map(|g| g.span.0)),
+            "the text {} could not be displayed with any font",
+            &text.text,
+        );
     }
 
-    let color_glyph_count =
-        text.glyphs.iter().filter(|g| is_color_glyph(&text.font, g)).count();
+    let outline_glyphs =
+        text.glyphs.iter().filter(|g| should_outline(&text.font, g)).count();
 
-    if color_glyph_count == text.glyphs.len() {
-        write_color_glyphs(ctx, pos, TextItemView::all_of(text));
-    } else if color_glyph_count == 0 {
-        write_normal_text(ctx, pos, TextItemView::all_of(text));
+    if outline_glyphs == text.glyphs.len() {
+        write_normal_text(ctx, pos, TextItemView::full(text))?;
+    } else if outline_glyphs == 0 {
+        write_complex_glyphs(ctx, pos, TextItemView::full(text))?;
     } else {
-        // Otherwise we need to split it in smaller text runs
+        // Otherwise we need to split it into smaller text runs.
         let mut offset = 0;
         let mut position_in_run = Abs::zero();
-        for (color, sub_run) in
-            text.glyphs.group_by_key(|g| is_color_glyph(&text.font, g))
+        for (should_outline, sub_run) in
+            text.glyphs.group_by_key(|g| should_outline(&text.font, g))
         {
             let end = offset + sub_run.len();
 
@@ -428,18 +449,25 @@ fn write_text(ctx: &mut Builder, pos: Point, text: &TextItem) {
             let pos = pos + Point::new(position_in_run, Abs::zero());
             position_in_run += text_item_view.width();
             offset = end;
-            // Actually write the sub text-run
-            if color {
-                write_color_glyphs(ctx, pos, text_item_view);
+
+            // Actually write the sub text-run.
+            if should_outline {
+                write_normal_text(ctx, pos, text_item_view)?;
             } else {
-                write_normal_text(ctx, pos, text_item_view);
+                write_complex_glyphs(ctx, pos, text_item_view)?;
             }
         }
     }
+
+    Ok(())
 }
 
 /// Encodes a text run (without any color glyph) into the content stream.
-fn write_normal_text(ctx: &mut Builder, pos: Point, text: TextItemView) {
+fn write_normal_text(
+    ctx: &mut Builder,
+    pos: Point,
+    text: TextItemView,
+) -> SourceResult<()> {
     let x = pos.x.to_f32();
     let y = pos.y.to_f32();
 
@@ -447,13 +475,11 @@ fn write_normal_text(ctx: &mut Builder, pos: Point, text: TextItemView) {
 
     let glyph_set = ctx.resources.glyph_sets.entry(text.item.font.clone()).or_default();
     for g in text.glyphs() {
-        let t = text.text();
-        let segment = &t[g.range()];
-        glyph_set.entry(g.id).or_insert_with(|| segment.into());
+        glyph_set.entry(g.id).or_insert_with(|| text.glyph_text(g));
     }
 
     let fill_transform = ctx.state.transforms(Size::zero(), pos);
-    ctx.set_fill(&text.item.fill, true, fill_transform);
+    ctx.set_fill(&text.item.fill, true, fill_transform)?;
 
     let stroke = text.item.stroke.as_ref().and_then(|stroke| {
         if stroke.thickness.to_f32() > 0.0 {
@@ -464,7 +490,7 @@ fn write_normal_text(ctx: &mut Builder, pos: Point, text: TextItemView) {
     });
 
     if let Some(stroke) = stroke {
-        ctx.set_stroke(stroke, true, fill_transform);
+        ctx.set_stroke(stroke, true, fill_transform)?;
         ctx.set_text_rendering_mode(TextRenderingMode::FillStroke);
     } else {
         ctx.set_text_rendering_mode(TextRenderingMode::Fill);
@@ -490,11 +516,15 @@ fn write_normal_text(ctx: &mut Builder, pos: Point, text: TextItemView) {
 
     // Write the glyphs with kerning adjustments.
     for glyph in text.glyphs() {
+        if ctx.options.standards.pdfa && glyph.id == 0 {
+            bail!(tofu(&text, glyph));
+        }
+
         adjustment += glyph.x_offset;
 
         if !adjustment.is_zero() {
             if !encoded.is_empty() {
-                items.show(Str(&encoded));
+                show_text(&mut items, &encoded);
                 encoded.clear();
             }
 
@@ -533,16 +563,30 @@ fn write_normal_text(ctx: &mut Builder, pos: Point, text: TextItemView) {
     }
 
     if !encoded.is_empty() {
-        items.show(Str(&encoded));
+        show_text(&mut items, &encoded);
     }
 
     items.finish();
     positioned.finish();
     ctx.content.end_text();
+
+    Ok(())
+}
+
+/// Shows text, ensuring that each individual string doesn't exceed the
+/// implementation limits.
+fn show_text(items: &mut PositionedItems, encoded: &[u8]) {
+    for chunk in encoded.chunks(Str::PDFA_LIMIT) {
+        items.show(Str(chunk));
+    }
 }
 
 /// Encodes a text run made only of color glyphs into the content stream
-fn write_color_glyphs(ctx: &mut Builder, pos: Point, text: TextItemView) {
+fn write_complex_glyphs(
+    ctx: &mut Builder,
+    pos: Point,
+    text: TextItemView,
+) -> SourceResult<()> {
     let x = pos.x.to_f32();
     let y = pos.y.to_f32();
 
@@ -563,12 +607,17 @@ fn write_color_glyphs(ctx: &mut Builder, pos: Point, text: TextItemView) {
         .or_default();
 
     for glyph in text.glyphs() {
+        if ctx.options.standards.pdfa && glyph.id == 0 {
+            bail!(tofu(&text, glyph));
+        }
+
         // Retrieve the Type3 font reference and the glyph index in the font.
         let color_fonts = ctx
             .resources
             .color_fonts
             .get_or_insert_with(|| Box::new(ColorFontMap::new()));
-        let (font, index) = color_fonts.get(&text.item.font, glyph.id);
+
+        let (font, index) = color_fonts.get(ctx.options, &text, glyph)?;
 
         if last_font != Some(font) {
             ctx.content.set_font(
@@ -580,15 +629,15 @@ fn write_color_glyphs(ctx: &mut Builder, pos: Point, text: TextItemView) {
 
         ctx.content.show(Str(&[index]));
 
-        glyph_set
-            .entry(glyph.id)
-            .or_insert_with(|| text.text()[glyph.range()].into());
+        glyph_set.entry(glyph.id).or_insert_with(|| text.glyph_text(glyph));
     }
     ctx.content.end_text();
+
+    Ok(())
 }
 
 /// Encode a geometrical shape into the content stream.
-fn write_shape(ctx: &mut Builder, pos: Point, shape: &Shape) {
+fn write_shape(ctx: &mut Builder, pos: Point, shape: &Shape) -> SourceResult<()> {
     let x = pos.x.to_f32();
     let y = pos.y.to_f32();
 
@@ -601,11 +650,11 @@ fn write_shape(ctx: &mut Builder, pos: Point, shape: &Shape) {
     });
 
     if shape.fill.is_none() && stroke.is_none() {
-        return;
+        return Ok(());
     }
 
     if let Some(fill) = &shape.fill {
-        ctx.set_fill(fill, false, ctx.state.transforms(shape.geometry.bbox_size(), pos));
+        ctx.set_fill(fill, false, ctx.state.transforms(shape.geometry.bbox_size(), pos))?;
     }
 
     if let Some(stroke) = stroke {
@@ -613,7 +662,7 @@ fn write_shape(ctx: &mut Builder, pos: Point, shape: &Shape) {
             stroke,
             false,
             ctx.state.transforms(shape.geometry.bbox_size(), pos),
-        );
+        )?;
     }
 
     ctx.set_opacities(stroke, shape.fill.as_ref());
@@ -645,6 +694,8 @@ fn write_shape(ctx: &mut Builder, pos: Point, shape: &Shape) {
         (Some(_), FillRule::NonZero, Some(_)) => ctx.content.fill_nonzero_and_stroke(),
         (Some(_), FillRule::EvenOdd, Some(_)) => ctx.content.fill_even_odd_and_stroke(),
     };
+
+    Ok(())
 }
 
 /// Encode a bezier path into the content stream.
@@ -671,14 +722,22 @@ fn write_path(ctx: &mut Builder, x: f32, y: f32, path: &Path) {
 }
 
 /// Encode a vector or raster image into the content stream.
-fn write_image(ctx: &mut Builder, x: f32, y: f32, image: &Image, size: Size) {
+fn write_image(
+    ctx: &mut Builder,
+    x: f32,
+    y: f32,
+    image: &Image,
+    size: Size,
+    span: Span,
+) -> SourceResult<()> {
     let index = ctx.resources.images.insert(image.clone());
     ctx.resources.deferred_images.entry(index).or_insert_with(|| {
-        let (image, color_space) = deferred_image(image.clone());
+        let (image, color_space) =
+            deferred_image(image.clone(), ctx.options.standards.pdfa);
         if let Some(color_space) = color_space {
             ctx.resources.colors.mark_as_used(color_space);
         }
-        image
+        (image, span)
     });
 
     ctx.reset_opacities();
@@ -686,14 +745,18 @@ fn write_image(ctx: &mut Builder, x: f32, y: f32, image: &Image, size: Size) {
     let name = eco_format!("Im{index}");
     let w = size.x.to_f32();
     let h = size.y.to_f32();
-    ctx.content.save_state();
+    ctx.content.save_state_checked()?;
     ctx.content.transform([w, 0.0, 0.0, -h, x, y + h]);
 
     if let Some(alt) = image.alt() {
+        if ctx.options.standards.pdfa && alt.len() > Str::PDFA_LIMIT {
+            bail!(span, "the image's alt text is too long");
+        }
+
         let mut image_span =
             ctx.content.begin_marked_content_with_properties(Name(b"Span"));
         let mut image_alt = image_span.properties();
-        image_alt.pair(Name(b"Alt"), pdf_writer::Str(alt.as_bytes()));
+        image_alt.pair(Name(b"Alt"), Str(alt.as_bytes()));
         image_alt.finish();
         image_span.finish();
 
@@ -704,6 +767,7 @@ fn write_image(ctx: &mut Builder, x: f32, y: f32, image: &Image, size: Size) {
     }
 
     ctx.content.restore_state();
+    Ok(())
 }
 
 /// Save a link for later writing in the annotations dictionary.
@@ -750,4 +814,14 @@ fn to_pdf_line_join(join: LineJoin) -> LineJoinStyle {
         LineJoin::Round => LineJoinStyle::RoundJoin,
         LineJoin::Bevel => LineJoinStyle::BevelJoin,
     }
+}
+
+/// The error when there is a tofu glyph.
+#[cold]
+fn tofu(text: &TextItemView, glyph: &Glyph) -> SourceDiagnostic {
+    error!(
+        glyph.span.0,
+        "the text {} could not be displayed with any font",
+        text.glyph_text(glyph).repr(),
+    )
 }
