@@ -15,17 +15,21 @@ mod pattern;
 mod resources;
 
 use std::collections::HashMap;
+use std::fmt::{self, Debug, Formatter};
 use std::hash::Hash;
 use std::ops::{Deref, DerefMut};
 
 use base64::Engine;
-use pdf_writer::{Chunk, Pdf, Ref};
-use typst::foundations::{Datetime, Smart};
-use typst::layout::{Abs, Em, PageRanges, Transform};
-use typst::model::Document;
-use typst::text::Font;
-use typst::utils::Deferred;
-use typst::visualize::Image;
+use pdf_writer::{Chunk, Name, Pdf, Ref, Str, TextStr};
+use serde::{Deserialize, Serialize};
+use typst_library::diag::{bail, SourceResult, StrResult};
+use typst_library::foundations::{Datetime, Smart};
+use typst_library::layout::{Abs, Em, PageRanges, Transform};
+use typst_library::model::Document;
+use typst_library::text::Font;
+use typst_library::visualize::Image;
+use typst_syntax::Span;
+use typst_utils::Deferred;
 
 use crate::catalog::write_catalog;
 use crate::color::{alloc_color_functions_refs, ColorFunctionRefs};
@@ -44,51 +48,100 @@ use crate::resources::{
 /// Export a document into a PDF file.
 ///
 /// Returns the raw bytes making up the PDF file.
-///
-/// The `ident` parameter, if given, shall be a string that uniquely and stably
-/// identifies the document. It should not change between compilations of the
-/// same document.  **If you cannot provide such a stable identifier, just pass
-/// `Smart::Auto` rather than trying to come up with one.** The CLI, for
-/// example, does not have a well-defined notion of a long-lived project and as
-/// such just passes `Smart::Auto`.
-///
-/// If an `ident` is given, the hash of it will be used to create a PDF document
-/// identifier (the identifier itself is not leaked). If `ident` is `Auto`, a
-/// hash of the document's title and author is used instead (which is reasonably
-/// unique and stable).
-///
-/// The `timestamp`, if given, is expected to be the creation date of the
-/// document as a UTC datetime. It will only be used if `set document(date: ..)`
-/// is `auto`.
-///
-/// The `page_ranges` option specifies which ranges of pages should be exported
-/// in the PDF. When `None`, all pages should be exported.
 #[typst_macros::time(name = "pdf")]
-pub fn pdf(
-    document: &Document,
-    ident: Smart<&str>,
-    timestamp: Option<Datetime>,
-    page_ranges: Option<PageRanges>,
-) -> Vec<u8> {
-    PdfBuilder::new(document, page_ranges)
-        .phase(|builder| builder.run(traverse_pages))
-        .phase(|builder| GlobalRefs {
-            color_functions: builder.run(alloc_color_functions_refs),
-            pages: builder.run(alloc_page_refs),
-            resources: builder.run(alloc_resources_refs),
-        })
-        .phase(|builder| References {
-            named_destinations: builder.run(write_named_destinations),
-            fonts: builder.run(write_fonts),
-            color_fonts: builder.run(write_color_fonts),
-            images: builder.run(write_images),
-            gradients: builder.run(write_gradients),
-            patterns: builder.run(write_patterns),
-            ext_gs: builder.run(write_graphic_states),
-        })
-        .phase(|builder| builder.run(write_page_tree))
-        .phase(|builder| builder.run(write_resource_dictionaries))
-        .export_with(ident, timestamp, write_catalog)
+pub fn pdf(document: &Document, options: &PdfOptions) -> SourceResult<Vec<u8>> {
+    PdfBuilder::new(document, options)
+        .phase(|builder| builder.run(traverse_pages))?
+        .phase(|builder| {
+            Ok(GlobalRefs {
+                color_functions: builder.run(alloc_color_functions_refs)?,
+                pages: builder.run(alloc_page_refs)?,
+                resources: builder.run(alloc_resources_refs)?,
+            })
+        })?
+        .phase(|builder| {
+            Ok(References {
+                named_destinations: builder.run(write_named_destinations)?,
+                fonts: builder.run(write_fonts)?,
+                color_fonts: builder.run(write_color_fonts)?,
+                images: builder.run(write_images)?,
+                gradients: builder.run(write_gradients)?,
+                patterns: builder.run(write_patterns)?,
+                ext_gs: builder.run(write_graphic_states)?,
+            })
+        })?
+        .phase(|builder| builder.run(write_page_tree))?
+        .phase(|builder| builder.run(write_resource_dictionaries))?
+        .export_with(write_catalog)
+}
+
+/// Settings for PDF export.
+#[derive(Debug, Default)]
+pub struct PdfOptions<'a> {
+    /// If not `Smart::Auto`, shall be a string that uniquely and stably
+    /// identifies the document. It should not change between compilations of
+    /// the same document.  **If you cannot provide such a stable identifier,
+    /// just pass `Smart::Auto` rather than trying to come up with one.** The
+    /// CLI, for example, does not have a well-defined notion of a long-lived
+    /// project and as such just passes `Smart::Auto`.
+    ///
+    /// If an `ident` is given, the hash of it will be used to create a PDF
+    /// document identifier (the identifier itself is not leaked). If `ident` is
+    /// `Auto`, a hash of the document's title and author is used instead (which
+    /// is reasonably unique and stable).
+    pub ident: Smart<&'a str>,
+    /// If not `None`, shall be the creation date of the document as a UTC
+    /// datetime. It will only be used if `set document(date: ..)` is `auto`.
+    pub timestamp: Option<Datetime>,
+    /// Specifies which ranges of pages should be exported in the PDF. When
+    /// `None`, all pages should be exported.
+    pub page_ranges: Option<PageRanges>,
+    /// A list of PDF standards that Typst will enforce conformance with.
+    pub standards: PdfStandards,
+}
+
+/// Encapsulates a list of compatible PDF standards.
+#[derive(Clone)]
+pub struct PdfStandards {
+    /// For now, we simplify to just PDF/A, since we only support PDF/A-2b. But
+    /// it can be more fine-grained in the future.
+    pub(crate) pdfa: bool,
+}
+
+impl PdfStandards {
+    /// Validates a list of PDF standards for compatibility and returns their
+    /// encapsulated representation.
+    pub fn new(list: &[PdfStandard]) -> StrResult<Self> {
+        Ok(Self { pdfa: list.contains(&PdfStandard::A_2b) })
+    }
+}
+
+impl Debug for PdfStandards {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        f.pad("PdfStandards(..)")
+    }
+}
+
+#[allow(clippy::derivable_impls)]
+impl Default for PdfStandards {
+    fn default() -> Self {
+        Self { pdfa: false }
+    }
+}
+
+/// A PDF standard that Typst can enforce conformance with.
+///
+/// Support for more standards is planned.
+#[derive(Debug, Copy, Clone, Eq, PartialEq, Serialize, Deserialize)]
+#[allow(non_camel_case_types)]
+#[non_exhaustive]
+pub enum PdfStandard {
+    /// PDF 1.7.
+    #[serde(rename = "1.7")]
+    V_1_7,
+    /// PDF/A-2b.
+    #[serde(rename = "a-2b")]
+    A_2b,
 }
 
 /// A struct to build a PDF following a fixed succession of phases.
@@ -124,9 +177,8 @@ struct PdfBuilder<S> {
 struct WithDocument<'a> {
     /// The Typst document that is exported.
     document: &'a Document,
-    /// Page ranges to export.
-    /// When `None`, all pages are exported.
-    exported_pages: Option<PageRanges>,
+    /// Settings for PDF export.
+    options: &'a PdfOptions<'a>,
 }
 
 /// At this point, resources were listed, but they don't have any reference
@@ -135,7 +187,7 @@ struct WithDocument<'a> {
 /// This phase allocates some global references.
 struct WithResources<'a> {
     document: &'a Document,
-    exported_pages: Option<PageRanges>,
+    options: &'a PdfOptions<'a>,
     /// The content of the pages encoded as PDF content streams.
     ///
     /// The pages are at the index corresponding to their page number, but they
@@ -170,7 +222,7 @@ impl<'a> From<(WithDocument<'a>, (Vec<Option<EncodedPage>>, Resources<()>))>
     ) -> Self {
         Self {
             document: previous.document,
-            exported_pages: previous.exported_pages,
+            options: previous.options,
             pages,
             resources,
         }
@@ -184,7 +236,7 @@ impl<'a> From<(WithDocument<'a>, (Vec<Option<EncodedPage>>, Resources<()>))>
 /// that will be collected in [`References`].
 struct WithGlobalRefs<'a> {
     document: &'a Document,
-    exported_pages: Option<PageRanges>,
+    options: &'a PdfOptions<'a>,
     pages: Vec<Option<EncodedPage>>,
     /// Resources are the same as in previous phases, but each dictionary now has a reference.
     resources: Resources,
@@ -196,7 +248,7 @@ impl<'a> From<(WithResources<'a>, GlobalRefs)> for WithGlobalRefs<'a> {
     fn from((previous, globals): (WithResources<'a>, GlobalRefs)) -> Self {
         Self {
             document: previous.document,
-            exported_pages: previous.exported_pages,
+            options: previous.options,
             pages: previous.pages,
             resources: previous.resources.with_refs(&globals.resources),
             globals,
@@ -226,10 +278,10 @@ struct References {
 /// tree is going to be written, and given a reference. It is also at this point that
 /// the page contents is actually written.
 struct WithRefs<'a> {
-    globals: GlobalRefs,
     document: &'a Document,
+    options: &'a PdfOptions<'a>,
+    globals: GlobalRefs,
     pages: Vec<Option<EncodedPage>>,
-    exported_pages: Option<PageRanges>,
     resources: Resources,
     /// References that were allocated for resources.
     references: References,
@@ -238,9 +290,9 @@ struct WithRefs<'a> {
 impl<'a> From<(WithGlobalRefs<'a>, References)> for WithRefs<'a> {
     fn from((previous, references): (WithGlobalRefs<'a>, References)) -> Self {
         Self {
-            globals: previous.globals,
-            exported_pages: previous.exported_pages,
             document: previous.document,
+            options: previous.options,
+            globals: previous.globals,
             pages: previous.pages,
             resources: previous.resources,
             references,
@@ -252,10 +304,10 @@ impl<'a> From<(WithGlobalRefs<'a>, References)> for WithRefs<'a> {
 ///
 /// Each sub-resource gets its own isolated resource dictionary.
 struct WithEverything<'a> {
-    globals: GlobalRefs,
     document: &'a Document,
+    options: &'a PdfOptions<'a>,
+    globals: GlobalRefs,
     pages: Vec<Option<EncodedPage>>,
-    exported_pages: Option<PageRanges>,
     resources: Resources,
     references: References,
     /// Reference that was allocated for the page tree.
@@ -271,9 +323,9 @@ impl<'a> From<(WithEverything<'a>, ())> for WithEverything<'a> {
 impl<'a> From<(WithRefs<'a>, Ref)> for WithEverything<'a> {
     fn from((previous, page_tree_ref): (WithRefs<'a>, Ref)) -> Self {
         Self {
-            exported_pages: previous.exported_pages,
-            globals: previous.globals,
             document: previous.document,
+            options: previous.options,
+            globals: previous.globals,
             resources: previous.resources,
             references: previous.references,
             pages: previous.pages,
@@ -284,42 +336,42 @@ impl<'a> From<(WithRefs<'a>, Ref)> for WithEverything<'a> {
 
 impl<'a> PdfBuilder<WithDocument<'a>> {
     /// Start building a PDF for a Typst document.
-    fn new(document: &'a Document, exported_pages: Option<PageRanges>) -> Self {
+    fn new(document: &'a Document, options: &'a PdfOptions<'a>) -> Self {
         Self {
             alloc: Ref::new(1),
             pdf: Pdf::new(),
-            state: WithDocument { document, exported_pages },
+            state: WithDocument { document, options },
         }
     }
 }
 
 impl<S> PdfBuilder<S> {
     /// Start a new phase, and save its output in the global state.
-    fn phase<NS, B, O>(mut self, builder: B) -> PdfBuilder<NS>
+    fn phase<NS, B, O>(mut self, builder: B) -> SourceResult<PdfBuilder<NS>>
     where
         // New state
         NS: From<(S, O)>,
         // Builder
-        B: Fn(&mut Self) -> O,
+        B: Fn(&mut Self) -> SourceResult<O>,
     {
-        let output = builder(&mut self);
-        PdfBuilder {
+        let output = builder(&mut self)?;
+        Ok(PdfBuilder {
             state: NS::from((self.state, output)),
             alloc: self.alloc,
             pdf: self.pdf,
-        }
+        })
     }
 
-    /// Runs a step with the current state, merge its output in the PDF file,
-    /// and renumber any references it returned.
-    fn run<P, O>(&mut self, process: P) -> O
+    /// Run a step with the current state, merges its output into the PDF file,
+    /// and renumbers any references it returned.
+    fn run<P, O>(&mut self, process: P) -> SourceResult<O>
     where
         // Process
-        P: Fn(&S) -> (PdfChunk, O),
+        P: Fn(&S) -> SourceResult<(PdfChunk, O)>,
         // Output
         O: Renumber,
     {
-        let (chunk, mut output) = process(&self.state);
+        let (chunk, mut output) = process(&self.state)?;
         // Allocate a final reference for each temporary one
         let allocated = chunk.alloc.get() - TEMPORARY_REFS_START;
         let offset = TEMPORARY_REFS_START - self.alloc.get();
@@ -336,22 +388,17 @@ impl<S> PdfBuilder<S> {
 
         self.alloc = Ref::new(self.alloc.get() + allocated);
 
-        output
+        Ok(output)
     }
 
     /// Finalize the PDF export and returns the buffer representing the
     /// document.
-    fn export_with<P>(
-        mut self,
-        ident: Smart<&str>,
-        timestamp: Option<Datetime>,
-        process: P,
-    ) -> Vec<u8>
+    fn export_with<P>(mut self, process: P) -> SourceResult<Vec<u8>>
     where
-        P: Fn(S, Smart<&str>, Option<Datetime>, &mut Pdf, &mut Ref),
+        P: Fn(S, &mut Pdf, &mut Ref) -> SourceResult<()>,
     {
-        process(self.state, ident, timestamp, &mut self.pdf, &mut self.alloc);
-        self.pdf.finish()
+        process(self.state, &mut self.pdf, &mut self.alloc)?;
+        Ok(self.pdf.finish())
     }
 }
 
@@ -417,7 +464,7 @@ struct PdfChunk {
 /// chunk, and should be remapped.
 ///
 /// This is a constant (large enough to avoid collisions) and not
-/// dependant on self.alloc to allow for better memoization of steps, if
+/// dependent on self.alloc to allow for better memoization of steps, if
 /// needed in the future.
 const TEMPORARY_REFS_START: i32 = 1_000_000_000;
 
@@ -471,7 +518,7 @@ fn deflate_deferred(content: Vec<u8>) -> Deferred<Vec<u8>> {
 /// Create a base64-encoded hash of the value.
 fn hash_base64<T: Hash>(value: &T) -> String {
     base64::engine::general_purpose::STANDARD
-        .encode(typst::utils::hash128(value).to_be_bytes())
+        .encode(typst_utils::hash128(value).to_be_bytes())
 }
 
 /// Additional methods for [`Abs`].
@@ -495,6 +542,63 @@ trait EmExt {
 impl EmExt for Em {
     fn to_font_units(self) -> f32 {
         1000.0 * self.get() as f32
+    }
+}
+
+trait NameExt<'a> {
+    /// The maximum length of a name in PDF/A.
+    const PDFA_LIMIT: usize = 127;
+}
+
+impl<'a> NameExt<'a> for Name<'a> {}
+
+/// Additional methods for [`Str`].
+trait StrExt<'a>: Sized {
+    /// The maximum length of a string in PDF/A.
+    const PDFA_LIMIT: usize = 32767;
+
+    /// Create a string that satisfies the constraints of PDF/A.
+    #[allow(unused)]
+    fn trimmed(string: &'a [u8]) -> Self;
+}
+
+impl<'a> StrExt<'a> for Str<'a> {
+    fn trimmed(string: &'a [u8]) -> Self {
+        Self(&string[..string.len().min(Self::PDFA_LIMIT)])
+    }
+}
+
+/// Additional methods for [`TextStr`].
+trait TextStrExt<'a>: Sized {
+    /// The maximum length of a string in PDF/A.
+    const PDFA_LIMIT: usize = Str::PDFA_LIMIT;
+
+    /// Create a text string that satisfies the constraints of PDF/A.
+    fn trimmed(string: &'a str) -> Self;
+}
+
+impl<'a> TextStrExt<'a> for TextStr<'a> {
+    fn trimmed(string: &'a str) -> Self {
+        Self(&string[..string.len().min(Self::PDFA_LIMIT)])
+    }
+}
+
+/// Extension trait for [`Content`](pdf_writer::Content).
+trait ContentExt {
+    fn save_state_checked(&mut self) -> SourceResult<()>;
+}
+
+impl ContentExt for pdf_writer::Content {
+    fn save_state_checked(&mut self) -> SourceResult<()> {
+        self.save_state();
+        if self.state_nesting_depth() > 28 {
+            bail!(
+                Span::detached(),
+                "maximum PDF grouping depth exceeding";
+                hint: "try to avoid excessive nesting of layout containers",
+            );
+        }
+        Ok(())
     }
 }
 
