@@ -13,71 +13,44 @@
 //!   order-independent and thus much better suited for further processing than
 //!   the raw markup.
 //! - **Layouting:**
-//!   Next, the content is [layouted] into a [document] containing one [frame]
+//!   Next, the content is [laid out] into a [document] containing one [frame]
 //!   per page with items at fixed positions.
 //! - **Exporting:**
 //!   These frames can finally be exported into an output format (currently PDF,
 //!   PNG, or SVG).
 //!
-//! [tokens]: syntax::SyntaxKind
-//! [parsed]: syntax::parse
-//! [syntax tree]: syntax::SyntaxNode
-//! [AST]: syntax::ast
-//! [evaluate]: eval::eval
-//! [module]: foundations::Module
-//! [content]: foundations::Content
-//! [layouted]: crate::layout::layout_document
-//! [document]: model::Document
-//! [frame]: layout::Frame
+//! [tokens]: typst_syntax::SyntaxKind
+//! [parsed]: typst_syntax::parse
+//! [syntax tree]: typst_syntax::SyntaxNode
+//! [AST]: typst_syntax::ast
+//! [evaluate]: typst_eval::eval
+//! [module]: crate::foundations::Module
+//! [content]: crate::foundations::Content
+//! [laid out]: typst_layout::layout_document
+//! [document]: crate::model::Document
+//! [frame]: crate::layout::Frame
 
-#![recursion_limit = "1000"]
-#![allow(clippy::comparison_chain)]
-#![allow(clippy::wildcard_in_or_patterns)]
-#![allow(clippy::manual_range_contains)]
+pub extern crate comemo;
+pub extern crate ecow;
 
-extern crate self as typst;
-
-pub mod diag;
-pub mod engine;
-pub mod eval;
-pub mod foundations;
-pub mod introspection;
-pub mod layout;
-pub mod loading;
-pub mod math;
-pub mod model;
-pub mod realize;
-pub mod symbols;
-pub mod text;
-pub mod visualize;
-
+pub use typst_library::*;
 #[doc(inline)]
 pub use typst_syntax as syntax;
 #[doc(inline)]
 pub use typst_utils as utils;
 
 use std::collections::HashSet;
-use std::ops::{Deref, Range};
 
 use comemo::{Track, Tracked, Validate};
 use ecow::{eco_format, eco_vec, EcoString, EcoVec};
+use typst_library::diag::{warning, FileError, SourceDiagnostic, SourceResult, Warned};
+use typst_library::engine::{Engine, Route, Sink, Traced};
+use typst_library::foundations::{StyleChain, Styles, Value};
+use typst_library::introspection::Introspector;
+use typst_library::model::Document;
+use typst_library::routines::Routines;
+use typst_syntax::{FileId, Span};
 use typst_timing::{timed, TimingScope};
-
-use crate::diag::{
-    warning, FileError, FileResult, SourceDiagnostic, SourceResult, Warned,
-};
-use crate::engine::{Engine, Route, Sink, Traced};
-use crate::foundations::{
-    Array, Bytes, Datetime, Dict, Module, Scope, StyleChain, Styles, Value,
-};
-use crate::introspection::Introspector;
-use crate::layout::{Alignment, Dir};
-use crate::model::Document;
-use crate::syntax::package::PackageSpec;
-use crate::syntax::{FileId, Source, Span};
-use crate::text::{Font, FontBook};
-use crate::utils::LazyHash;
-use crate::visualize::Color;
 
 /// Compile sources into a fully layouted document.
 ///
@@ -118,7 +91,8 @@ fn compile_impl(
         .map_err(|err| hint_invalid_main_file(world, err, main))?;
 
     // First evaluate the main source file into a module.
-    let content = crate::eval::eval(
+    let content = typst_eval::eval(
+        &ROUTINES,
         world,
         traced,
         sink.track_mut(),
@@ -128,6 +102,7 @@ fn compile_impl(
     .content();
 
     let mut iter = 0;
+    let mut subsink;
     let mut document = Document::default();
 
     // Relayout until all introspections stabilize.
@@ -136,23 +111,22 @@ fn compile_impl(
         // The name of the iterations for timing scopes.
         const ITER_NAMES: &[&str] =
             &["layout (1)", "layout (2)", "layout (3)", "layout (4)", "layout (5)"];
-        let _scope = TimingScope::new(ITER_NAMES[iter], None);
+        let _scope = TimingScope::new(ITER_NAMES[iter]);
 
-        // Clear delayed errors.
-        sink.delayed();
+        subsink = Sink::new();
 
         let constraint = <Introspector as Validate>::Constraint::new();
         let mut engine = Engine {
             world,
             introspector: document.introspector.track_with(&constraint),
             traced,
-            sink: sink.track_mut(),
+            sink: subsink.track_mut(),
             route: Route::default(),
+            routines: &ROUTINES,
         };
 
         // Layout!
-        document = crate::layout::layout_document(&mut engine, &content, styles)?;
-        document.introspector.rebuild(&document.pages);
+        document = (engine.routines.layout_document)(&mut engine, &content, styles)?;
         iter += 1;
 
         if timed!("check stabilized", document.introspector.validate(&constraint)) {
@@ -160,13 +134,15 @@ fn compile_impl(
         }
 
         if iter >= 5 {
-            sink.warn(warning!(
+            subsink.warn(warning!(
                 Span::detached(), "layout did not converge within 5 attempts";
                 hint: "check if any states or queries are updating themselves"
             ));
             break;
         }
     }
+
+    sink.extend_from_sink(subsink);
 
     // Promote delayed errors.
     let delayed = sink.delayed();
@@ -181,235 +157,10 @@ fn compile_impl(
 fn deduplicate(mut diags: EcoVec<SourceDiagnostic>) -> EcoVec<SourceDiagnostic> {
     let mut unique = HashSet::new();
     diags.retain(|diag| {
-        let hash = crate::utils::hash128(&(&diag.span, &diag.message));
+        let hash = typst_utils::hash128(&(&diag.span, &diag.message));
         unique.insert(hash)
     });
     diags
-}
-
-/// The environment in which typesetting occurs.
-///
-/// All loading functions (`main`, `source`, `file`, `font`) should perform
-/// internal caching so that they are relatively cheap on repeated invocations
-/// with the same argument. [`Source`], [`Bytes`], and [`Font`] are
-/// all reference-counted and thus cheap to clone.
-///
-/// The compiler doesn't do the caching itself because the world has much more
-/// information on when something can change. For example, fonts typically don't
-/// change and can thus even be cached across multiple compilations (for
-/// long-running applications like `typst watch`). Source files on the other
-/// hand can change and should thus be cleared after each compilation. Advanced
-/// clients like language servers can also retain the source files and
-/// [edit](Source::edit) them in-place to benefit from better incremental
-/// performance.
-#[comemo::track]
-pub trait World: Send + Sync {
-    /// The standard library.
-    ///
-    /// Can be created through `Library::build()`.
-    fn library(&self) -> &LazyHash<Library>;
-
-    /// Metadata about all known fonts.
-    fn book(&self) -> &LazyHash<FontBook>;
-
-    /// Get the file id of the main source file.
-    fn main(&self) -> FileId;
-
-    /// Try to access the specified source file.
-    fn source(&self, id: FileId) -> FileResult<Source>;
-
-    /// Try to access the specified file.
-    fn file(&self, id: FileId) -> FileResult<Bytes>;
-
-    /// Try to access the font with the given index in the font book.
-    fn font(&self, index: usize) -> Option<Font>;
-
-    /// Get the current date.
-    ///
-    /// If no offset is specified, the local date should be chosen. Otherwise,
-    /// the UTC date should be chosen with the corresponding offset in hours.
-    ///
-    /// If this function returns `None`, Typst's `datetime` function will
-    /// return an error.
-    fn today(&self, offset: Option<i64>) -> Option<Datetime>;
-
-    /// A list of all available packages and optionally descriptions for them.
-    ///
-    /// This function is optional to implement. It enhances the user experience
-    /// by enabling autocompletion for packages. Details about packages from the
-    /// `@preview` namespace are available from
-    /// `https://packages.typst.org/preview/index.json`.
-    fn packages(&self) -> &[(PackageSpec, Option<EcoString>)] {
-        &[]
-    }
-}
-
-macro_rules! delegate_for_ptr {
-    ($W:ident for $ptr:ty) => {
-        impl<$W: World> World for $ptr {
-            fn library(&self) -> &LazyHash<Library> {
-                self.deref().library()
-            }
-
-            fn book(&self) -> &LazyHash<FontBook> {
-                self.deref().book()
-            }
-
-            fn main(&self) -> FileId {
-                self.deref().main()
-            }
-
-            fn source(&self, id: FileId) -> FileResult<Source> {
-                self.deref().source(id)
-            }
-
-            fn file(&self, id: FileId) -> FileResult<Bytes> {
-                self.deref().file(id)
-            }
-
-            fn font(&self, index: usize) -> Option<Font> {
-                self.deref().font(index)
-            }
-
-            fn today(&self, offset: Option<i64>) -> Option<Datetime> {
-                self.deref().today(offset)
-            }
-
-            fn packages(&self) -> &[(PackageSpec, Option<EcoString>)] {
-                self.deref().packages()
-            }
-        }
-    };
-}
-
-delegate_for_ptr!(W for std::boxed::Box<W>);
-delegate_for_ptr!(W for std::sync::Arc<W>);
-delegate_for_ptr!(W for &W);
-
-/// Helper methods on [`World`] implementations.
-pub trait WorldExt {
-    /// Get the byte range for a span.
-    ///
-    /// Returns `None` if the `Span` does not point into any source file.
-    fn range(&self, span: Span) -> Option<Range<usize>>;
-}
-
-impl<T: World> WorldExt for T {
-    fn range(&self, span: Span) -> Option<Range<usize>> {
-        self.source(span.id()?).ok()?.range(span)
-    }
-}
-
-/// Definition of Typst's standard library.
-#[derive(Debug, Clone, Hash)]
-pub struct Library {
-    /// The module that contains the definitions that are available everywhere.
-    pub global: Module,
-    /// The module that contains the definitions available in math mode.
-    pub math: Module,
-    /// The default style properties (for page size, font selection, and
-    /// everything else configurable via set and show rules).
-    pub styles: Styles,
-    /// The standard library as a value.
-    /// Used to provide the `std` variable.
-    pub std: Value,
-}
-
-impl Library {
-    /// Create a new builder for a library.
-    pub fn builder() -> LibraryBuilder {
-        LibraryBuilder::default()
-    }
-}
-
-impl Default for Library {
-    /// Constructs the standard library with the default configuration.
-    fn default() -> Self {
-        Self::builder().build()
-    }
-}
-
-/// Configurable builder for the standard library.
-///
-/// This struct is created by [`Library::builder`].
-#[derive(Debug, Clone, Default)]
-pub struct LibraryBuilder {
-    inputs: Option<Dict>,
-}
-
-impl LibraryBuilder {
-    /// Configure the inputs visible through `sys.inputs`.
-    pub fn with_inputs(mut self, inputs: Dict) -> Self {
-        self.inputs = Some(inputs);
-        self
-    }
-
-    /// Consumes the builder and returns a `Library`.
-    pub fn build(self) -> Library {
-        let math = math::module();
-        let inputs = self.inputs.unwrap_or_default();
-        let global = global(math.clone(), inputs);
-        let std = Value::Module(global.clone());
-        Library { global, math, styles: Styles::new(), std }
-    }
-}
-
-/// Construct the module with global definitions.
-fn global(math: Module, inputs: Dict) -> Module {
-    let mut global = Scope::deduplicating();
-    self::foundations::define(&mut global, inputs);
-    self::model::define(&mut global);
-    self::text::define(&mut global);
-    global.reset_category();
-    global.define_module(math);
-    self::layout::define(&mut global);
-    self::visualize::define(&mut global);
-    self::introspection::define(&mut global);
-    self::loading::define(&mut global);
-    self::symbols::define(&mut global);
-    prelude(&mut global);
-    Module::new("global", global)
-}
-
-/// Defines scoped values that are globally available, too.
-fn prelude(global: &mut Scope) {
-    global.reset_category();
-    global.define("black", Color::BLACK);
-    global.define("gray", Color::GRAY);
-    global.define("silver", Color::SILVER);
-    global.define("white", Color::WHITE);
-    global.define("navy", Color::NAVY);
-    global.define("blue", Color::BLUE);
-    global.define("aqua", Color::AQUA);
-    global.define("teal", Color::TEAL);
-    global.define("eastern", Color::EASTERN);
-    global.define("purple", Color::PURPLE);
-    global.define("fuchsia", Color::FUCHSIA);
-    global.define("maroon", Color::MAROON);
-    global.define("red", Color::RED);
-    global.define("orange", Color::ORANGE);
-    global.define("yellow", Color::YELLOW);
-    global.define("olive", Color::OLIVE);
-    global.define("green", Color::GREEN);
-    global.define("lime", Color::LIME);
-    global.define("luma", Color::luma_data());
-    global.define("oklab", Color::oklab_data());
-    global.define("oklch", Color::oklch_data());
-    global.define("rgb", Color::rgb_data());
-    global.define("cmyk", Color::cmyk_data());
-    global.define("range", Array::range_data());
-    global.define("ltr", Dir::LTR);
-    global.define("rtl", Dir::RTL);
-    global.define("ttb", Dir::TTB);
-    global.define("btt", Dir::BTT);
-    global.define("start", Alignment::START);
-    global.define("left", Alignment::LEFT);
-    global.define("center", Alignment::CENTER);
-    global.define("right", Alignment::RIGHT);
-    global.define("end", Alignment::END);
-    global.define("top", Alignment::TOP);
-    global.define("horizon", Alignment::HORIZON);
-    global.define("bottom", Alignment::BOTTOM);
 }
 
 /// Adds useful hints when the main source file couldn't be read
@@ -456,3 +207,40 @@ fn hint_invalid_main_file(
 
     eco_vec![diagnostic]
 }
+
+/// Defines implementation of various Typst compiler routines as a table of
+/// function pointers.
+///
+/// This is essentially dynamic linking and done to allow for crate splitting.
+pub static ROUTINES: Routines = Routines {
+    eval_string: typst_eval::eval_string,
+    eval_closure: typst_eval::eval_closure,
+    realize: typst_realize::realize,
+    layout_document: typst_layout::layout_document,
+    layout_fragment: typst_layout::layout_fragment,
+    layout_frame: typst_layout::layout_frame,
+    layout_inline: typst_layout::layout_inline,
+    layout_box: typst_layout::layout_box,
+    layout_list: typst_layout::layout_list,
+    layout_enum: typst_layout::layout_enum,
+    layout_grid: typst_layout::layout_grid,
+    layout_table: typst_layout::layout_table,
+    layout_stack: typst_layout::layout_stack,
+    layout_columns: typst_layout::layout_columns,
+    layout_move: typst_layout::layout_move,
+    layout_rotate: typst_layout::layout_rotate,
+    layout_scale: typst_layout::layout_scale,
+    layout_skew: typst_layout::layout_skew,
+    layout_repeat: typst_layout::layout_repeat,
+    layout_pad: typst_layout::layout_pad,
+    layout_line: typst_layout::layout_line,
+    layout_path: typst_layout::layout_path,
+    layout_polygon: typst_layout::layout_polygon,
+    layout_rect: typst_layout::layout_rect,
+    layout_square: typst_layout::layout_square,
+    layout_ellipse: typst_layout::layout_ellipse,
+    layout_circle: typst_layout::layout_circle,
+    layout_image: typst_layout::layout_image,
+    layout_equation_block: typst_layout::layout_equation_block,
+    layout_equation_inline: typst_layout::layout_equation_inline,
+};
